backend/requirements.txt
fastapi==0.110.0
uvicorn==0.23.0
python-jose==3.3.0
msal==1.23.0
httpx==0.24.0
sqlalchemy==1.4.53
asyncpg==0.28.0
alembic==1.11.1
pydantic==1.10.9
httpx==0.24.0
aiohttp==3.8.4
requests==2.31.0
pytest==7.4.0
pytest-asyncio==0.21.0
psycopg2-binary==2.9.6
redis==4.5.3
python-dotenv==1.1.0
